# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Agent System",
    "description": "企业级多Agent知识库助手系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat/pause": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Pause Stream",
        "description": "暂停正在执行的流式对话。",
        "operationId": "pause_stream_api_v1_chat_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_PauseStreamResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/ask": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Ask Endpoint",
        "operationId": "ask_endpoint_api_v1_chat_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response or SSE stream",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_AskResponse_"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "stream": {
                    "summary": "SSE stream",
                    "value": "event: content\ndata: {\"type\":\"content\",\"content\":\"hello\"}\n\n"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unified error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-sse-event-schema": {
          "$ref": "#/components/schemas/SSEEvent"
        },
        "x-stream-response": true
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Question",
            "description": "鐢ㄦ埛闂"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "浼氳瘽ID锛堝彲閫夛紝涓嶆彁渚涘垯鍒涘缓鏂颁細璇濓級"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "鏄惁浣跨敤娴佸紡杈撳嚭",
            "default": true
          },
          "enable_knowledge_base": {
            "type": "boolean",
            "title": "Enable Knowledge Base",
            "description": "是否启用知识库增强",
            "default": false
          },
          "knowledge_base_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Base Id",
            "description": "Selected knowledge base ID"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "description": "鐢ㄦ埛鎻愰棶璇锋眰"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseRequest": {
        "properties": {
          "stream_id": {
            "type": "string",
            "minLength": 1,
            "title": "Stream Id",
            "description": "娴佸紡浼氳瘽ID"
          }
        },
        "type": "object",
        "required": [
          "stream_id"
        ],
        "title": "PauseRequest",
        "description": "鏆傚仠娴佸紡瀵硅瘽璇锋眰"
      },
      "PauseStreamResponse": {
        "properties": {
          "stream_id": {
            "type": "string",
            "title": "Stream Id",
            "description": "???? ID"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "description": "?????"
          }
        },
        "type": "object",
        "required": [
          "stream_id",
          "paused"
        ],
        "title": "PauseStreamResponse"
      },
...
      "SuccessResponse_PauseStreamResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PauseStreamResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "SuccessResponse[PauseStreamResponse]",
        "example": {
          "code": 200,
          "data": {
            "key": "value"
          },
          "message": "success",
          "timestamp": "2024-01-01T00:00:00+00:00"
        }
      },
...
      "ErrorResponse": {
        "example": {
          "code": 400,
          "details": [
            {
              "field": "email",
              "message": "Invalid email",
              "type": "value_error"
            }
          ],
          "error": "ValidationError",
          "error_code": "SYSTEM_VALIDATION_ERROR",
          "message": "Validation error",
          "timestamp": "2024-01-01T00:00:00+00:00"
        },
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "error",
          "error_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SuccessResponse_AskResponse_": {
        "example": {
          "code": 200,
          "data": {
            "key": "value"
          },
          "message": "success",
          "timestamp": "2024-01-01T00:00:00+00:00"
        },
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AskResponse"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "title": "SuccessResponse[AskResponse]",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}